+++ app/toTwig/Converter/OxcontentConverter.php
@@ -12,7 +12,7 @@ namespace toTwig\Converter;
 class OxcontentConverter extends ConverterAbstract
 {
     protected string $name = 'oxcontent';
-    protected string $description = "Convert smarty {oxcontent} to twig function {% include 'content::...' %}";
+    protected string $description = "Convert smarty {oxcontent} to twig tag {% include_content %}";
     protected int $priority = 100;
 
     public function convert(string $content): string
@@ -24,40 +24,14 @@ class OxcontentConverter extends ConverterAbstract
             function ($matches) {
                 $attributes = $this->getAttributes($matches);
 
-                $key = null;
-                if (isset($attributes['ident'])) {
-                    $key = 'ident';
-                } elseif (isset($attributes['oxid'])) {
-                    $key = 'oxid';
-                }
-
-                $value = $this->rawString($this->sanitizeValue($attributes[$key]));
-
-                $templateName = "content::$key::$value";
-
-                $parameters = array_map(
-                    function ($attribute) {
-                        return $this->rawString($this->sanitizeValue($attribute));
-                    },
-                    array_filter(
-                        $attributes,
-                        function ($attribute) {
-                            return !in_array($attribute, ['ident', 'oxid', 'assign']);
-                        },
-                        ARRAY_FILTER_USE_KEY
-                    )
-                );
-
-                if (!empty($parameters)) {
-                    $templateName .= '?' . http_build_query($parameters);
-                }
+                $value = $this->rawString($this->sanitizeValue($attributes['ident']));
 
                 // Different approaches for syntax with and without assignment
                 $assignVar = $this->extractAssignVariableName($attributes);
                 if ($assignVar) {
-                    $twigTag = "{% set $assignVar = include('$templateName') %}";
+                    $twigTag = "{% set $assignVar = content('$value') %}";
                 } else {
-                    $twigTag = "{% include '$templateName' %}";
+                    $twigTag = "{% include_content '$value' %}";
                 }
 
                 return $twigTag;